Refactor PermissionGuard and add system overview documentation
